optimize the (used-by-PM) bulb functions, also fix a bug in (unused) bulb_emission_to_filament_temperature()

+++ src/wpc/bulb.h
@@ -14,11 +14,11 @@
 #define BULB_T_MAX 3400
 
 extern void bulb_init();
-extern double bulb_filament_temperature_to_emission(const double T);
+extern float bulb_filament_temperature_to_emission(const float T);
 extern double bulb_emission_to_filament_temperature(const double p);
 extern double bulb_cool_down_factor(const int bulb, const double T);
 extern double bulb_cool_down(const int bulb, double T, double duration);
-extern double bulb_heat_up_factor(const int bulb, const double T, const double U, const double serial_R);
+extern float bulb_heat_up_factor(const int bulb, const float T, const float U, const float serial_R);
 extern double bulb_heat_up(const int bulb, double T, double duration, const double U, const double serial_R);
 
 #endif /* INC_BULB */
